Moving adding listeners to Editor component.
