+++ libs/common-components/src/lib/document/document.component.ts
@@ -5,10 +5,7 @@ import {
   Input,
   OnInit,
   AfterViewInit,
-  OnChanges,
-  Output,
-  EventEmitter,
-  Renderer2
+  OnChanges
 } from '@angular/core';
 import {FileDescription, FileUtil} from "../file.service";
 import {ZoomService} from "../zoom.service";
@@ -51,16 +48,10 @@ export class DocumentComponent implements OnInit, AfterViewChecked, AfterViewIni
   curHeight = 0;
   isDesktop: boolean;
 
-  renderer: Renderer2;
-
   constructor(protected _elementRef: ElementRef<HTMLElement>,
               private _zoomService: ZoomService,
               private _windowService: WindowService,
-              private _navigateService: NavigateService,
-              renderer: Renderer2) {
-
-     this.renderer = renderer;
-
+              private _navigateService: NavigateService) {
     _zoomService.zoomChange.subscribe((val: number) => {
       this.zoom = val;
     });
@@ -102,30 +93,6 @@ export class DocumentComponent implements OnInit, AfterViewChecked, AfterViewIni
     this.curHeight = this.docHeight * this.scale;
 
     const hammer = new Hammer(this.container);
-
-    this.initControlsListeners();
-  }
-
-  private initControlsListeners() {
-    const inputs = this._elementRef.nativeElement.querySelectorAll('input');
-    inputs.forEach(input => {
-      this.renderer.listen(input, 'keyup', (event) => {
-        input.setAttribute('value', input.value);
-      });
-    });
-
-    const selects = this._elementRef.nativeElement.querySelectorAll('select');
-    selects.forEach(select => {
-      this.renderer.listen(select, 'change', (event) => {
-        selects.forEach(s => {
-          for (let i = s.options.length - 1; i >= 0; i--) {
-            s.options[i].removeAttribute('selected');
-          }
-        });
-
-        select.options[select.selectedIndex].setAttribute('selected', 'selected');
-      });
-    });
   }
 
   // TODO: this temporary crutch for Excel files should be documented
