+++ libs/editor/src/lib/editor-app.component.ts
@@ -1,4 +1,4 @@
-import {Component, AfterViewInit, OnInit} from '@angular/core';
+import {Component, AfterViewInit, OnInit, Renderer2} from '@angular/core';
 import {EditorService} from "./editor.service";
 import {
   FileDescription,
@@ -30,7 +30,7 @@ const $ = jquery;
   templateUrl: './editor-app.component.html',
   styleUrls: ['./editor-app.component.less']
 })
-export class EditorAppComponent implements OnInit, AfterViewInit  {
+export class EditorAppComponent implements OnInit, AfterViewInit {
   title = 'editor';
   files: FileModel[] = [];
   file: FileDescription;
@@ -67,6 +67,7 @@ export class EditorAppComponent implements OnInit, AfterViewInit  {
               private _htmlService: EditHtmlService,
               private _renderPrintService: RenderPrintService,
               private _loadingMaskService: LoadingMaskService,
+              private _renderer: Renderer2
   ) {
     this.isIE = /*@cc_on!@*/false || !!/(MSIE|Trident\/|Edge\/)/i.test(navigator.userAgent);
     configService.updatedConfig.subscribe((editorConfig) => {
@@ -292,6 +293,38 @@ export class EditorAppComponent implements OnInit, AfterViewInit  {
     }
     this.formatDisabled = !this.file;
     this.downloadDisabled = false;
+
+    // adding listeners on inputs if present on existing page
+    const timerId = setInterval(() => { 
+      let page = document.querySelectorAll('.page');
+      if (page)
+      {
+        this.initControlsListeners();
+        clearInterval(timerId);
+      }
+    }, 1000);
+  }
+
+  private initControlsListeners() {
+    const inputs = document.querySelectorAll('input');
+    inputs.forEach(input => {
+      this._renderer.listen(input, 'keyup', (event) => {
+        input.setAttribute('value', input.value);
+      });
+    });
+
+    const selects = document.querySelectorAll('select');
+    selects.forEach(select => {
+      this._renderer.listen(select, 'change', (event) => {
+        selects.forEach(s => {
+          for (let i = s.options.length - 1; i >= 0; i--) {
+            s.options[i].removeAttribute('selected');
+          }
+        });
+
+        select.options[select.selectedIndex].setAttribute('selected', 'selected');
+      });
+    });
   }
 
   private clearData() {
@@ -451,6 +484,11 @@ export class EditorAppComponent implements OnInit, AfterViewInit  {
     this.colorPickerShow = false;
     this.bgColorPickerShow = false;
     this._onCloseService.close(true);
+
+    // we try to save the changes on any click outside
+    if (document.querySelectorAll('.gd-wrapper')[0]) {
+      this.textBackup = document.querySelectorAll('.gd-wrapper')[0].innerHTML.toString();
+    }
   }
 
   toggleStrikeout(event) {
